♻️ refactor: getIp를 common/util로 통합, logger interceptor 중복 로직 제거
