+++ server/src/common/util/getIp.ts
@@ -1,0 +1,12 @@
+import { Request } from 'express';
+
+export function getIp(request: Request) {
+  const forwardedFor = request.headers['x-forwarded-for'];
+
+  if (typeof forwardedFor === 'string') {
+    const forwardedIps = forwardedFor.split(',');
+    return forwardedIps[0].trim();
+  }
+
+  return request.socket.remoteAddress;
+}
