+++ server/src/feed/util/viewCookie.ts
@@ -1,20 +1,9 @@
-import { Request, Response } from 'express';
+import { Response } from 'express';
 
 import { cookieConfig } from '@common/cookie/cookie.config';
 
 export function isString(ip: string | string[]): ip is string {
   return !Array.isArray(ip);
-}
-
-export function getIp(request: Request) {
-  const forwardedFor = request.headers['x-forwarded-for'];
-
-  if (typeof forwardedFor === 'string') {
-    const forwardedIps = forwardedFor.split(',');
-    return forwardedIps[0].trim();
-  }
-
-  return request.socket.remoteAddress;
 }
 
 function getExpirationTime() {
